% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIEXP3.R
\name{CIEXP3}
\alias{CIEXP3}
\title{Gives estimated values of SSI and its confidence bounds}
\usage{
CIEXP3(x1, x2, x3, alpha)
}
\arguments{
\item{x1}{is sample from Exp Population 1.}

\item{x2}{is sample from Exp Population 2.}

\item{x3}{is sample from Exp Population 3.}

\item{alpha}{is the level of significance}
}
\description{
Gives estimated values of SSI and its confidence bounds
}
